refactor(core): move validateDefinition into @bakin/core/workflows
Extract validateDefinition, its ValidateDefinitionOptions type, and the
private helpers only it uses (collectStepIds, getTopLevelStepIndex,
isSymbolicAgent, resolvePreferredForValidation, plus the exported
parsePreferredAgentExpression) from plugins/workflows/lib/parser.ts into
packages/core/src/workflows/validate-definition.ts — bodies byte-for-byte.
The agent-package loader and package-integrity checks now import it from
core instead of the plugin's parser; parser.ts re-exports it from the
core home so plugin-internal consumers (routes, start-validation,
load-defaults, step-context) are untouched. loadDefinition/listDefinitions
and the YAML/traversal helpers stay in the plugin.
